src/test/java/interviews/uber/kPalindrome/SolutionITest.java
package interviews.uber.kPalindrome;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertTrue;

/**
 * @author jacka
 * @version 1.0 on 12/11/2019
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase1(Solution solution) {
    final String input = "abc";
    final int k = 2;
    assertFalse(solution.canFormKPalindrome(input, k));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase2(Solution solution) {
    final String input = "aba";
    final int k = 2;
    assertTrue(solution.canFormKPalindrome(input, k));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase3(Solution solution) {
    final String input = "abc";
    final int k = 4;
    assertFalse(solution.canFormKPalindrome(input, k));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase4(Solution solution) {
    final String input = "abc";
    final int k = 3;
    assertTrue(solution.canFormKPalindrome(input, k));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase5(Solution solution) {
    final String input = "";
    final int k = 1;
    assertFalse(solution.canFormKPalindrome(input, k));
  }

  @ParameterizedTest
  @MethodSource("solutionProvider")
  void testOnlineCase6(Solution solution) {
    final String input = "";
    final int k = 0;
    assertTrue(solution.canFormKPalindrome(input, k));
  }

  static Stream<Solution> solutionProvider() {
    return Stream.of(
        new SolutionI()
    );
  }
}